mail-api/image/requirements.txt
kafka-python==2.0.2
Flask==2.0.1
Flask-Cors==3.0.10